Move topic map and peer map and gossip buffer into separate files
