+++ p2panda-net/src/lib.rs
@@ -10,6 +10,7 @@ mod sync;
 
 pub use addrs::{NodeAddress, RelayUrl};
 pub use config::Config;
+use iroh_net::key::PublicKey;
 pub use message::{FromBytes, ToBytes};
 pub use network::{Network, NetworkBuilder, RelayMode};
 pub use protocols::ProtocolHandler;
@@ -27,4 +28,8 @@ pub type NetworkId = [u8; 32];
 /// to synchronize past state before entering "live mode".
 pub trait TopicId {
     fn id(&self) -> [u8; 32];
+}
+
+pub(crate) fn to_public_key(key: PublicKey) -> p2panda_core::PublicKey {
+    p2panda_core::PublicKey::from_bytes(key.as_bytes()).expect("already validated public key")
 }
